+++ tests/testthat/test-to_random_case.R
@@ -1,0 +1,6 @@
+test_that("random_case", {
+  expect_equal(
+    {set.seed(123); to_random_case("almost RANDOM")},
+    "AlMosT raNdOm"
+  )
+})